+++ static_strings_for_data_extraction.py
